added separate classes for pickupable skill and heal
