--- Assets/Prefabs/Entities/Items/TestItem2.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5a03b8b12cfe3b645b38a63a8c61d33b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
