--- Assets/Prefabs/Entities/Items/TestItem3.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 173234e57878226479a35a0ec9061d0c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
