+++ Assets/Scripts/Items/PickupableHeal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c1da48366553e984ebf2fb22cc3644ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
