+++ Assets/Scripts/Items/PickupableSkill.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04fecee839d8a4a4c9156671d476ddb1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
